requirements.txt
# RAG Pipeline Dependencies

# PDF processing
PyMuPDF==1.24.0

# LLM API
groq==0.9.0

# Environment variables
python-dotenv==1.0.0

# Embeddings
sentence-transformers==2.5.1

# Vector database
chromadb==0.4.24

# Console output formatting
rich==13.7.0

